pyproject.toml
[project]
name = "housing"
version = "0.1.0"
description = "Project about Pipelines"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "joblib>=1.5.3",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
[tool.ruff]
line-length = 88  # Longitud estándar (evita líneas infinitas)
target-version = "py313" # Adaptado a tu versión de Python

[tool.ruff.lint]
# E: Errores de estilo (espacios, etc.)
# F: Errores lógicos (variables no usadas)
# I: ISORT (¡Importante!) Ordena tus imports alfabéticamente
# B: Bugbear (encuentra bugs comunes)
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double" # Estandariza comillas dobles
